Display generic models on organisation admin (#996)

+++ src/Application/VisaModel/Command/DeleteVisaModelCommandHandler.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Application\VisaModel\Command;
 
+use App\Domain\VisaModel\Exception\VisaModelCannotBeDeletedException;
 use App\Domain\VisaModel\Exception\VisaModelNotFoundException;
 use App\Domain\VisaModel\Repository\VisaModelRepositoryInterface;
 use App\Domain\VisaModel\VisaModel;
@@ -20,6 +21,10 @@ final class DeleteVisaModelCommandHandler
         $visaModel = $this->visaModelRepository->findOneByUuid($command->uuid);
         if (!$visaModel instanceof VisaModel) {
             throw new VisaModelNotFoundException();
+        }
+
+        if (!$visaModel->getOrganization()) {
+            throw new VisaModelCannotBeDeletedException();
         }
 
         $this->visaModelRepository->remove($visaModel);
